Unify producer + signature overlays into one incremental update
Fold the /Producer metadata stamp and the AcroForm signature injection into
a single appended revision instead of two chained incremental updates. The
shared xref/trailer serialization now lives in `pdf_scan` as
`append_incremental_update` (taking a list of `UpdatedObject`s), and the old
`sig_overlay` + `meta_overlay` modules collapse into one `overlay` module.

Benefits:
- One appended xref/trailer and one `/Prev` link instead of two.
- The PDF is parsed once, not once per pass.
- `/Info` and `/ID` are forwarded a single time, removing the cross-pass
  trailer-key bookkeeping the two-pass version required.
- Object ids come from one counter, so the (defensive) fresh `/Info` object
  can't collide with signature widget ids.

Also tightens comments around the non-obvious PDF mechanics (xref subsection
grouping, the two-part AcroForm/Annots requirement, object-token separation).
