+++ crates/backends/typst/src/overlay/extract.rs
